+++ clang/tools/libclang/CXLoadedDiagnostic.cpp
@@ -20,7 +20,7 @@
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/ADT/Twine.h"
-#include "llvm/Bitcode/BitstreamReader.h"
+#include "llvm/Bitstream/BitstreamReader.h"
 #include "llvm/Support/ErrorHandling.h"
 
 using namespace clang;
