+++ llvm/lib/Bitstream/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(Reader)
+# The writer is header-only.
