+++ llvm/lib/Bitstream/Reader/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_llvm_library(LLVMBitstreamReader
+  BitstreamReader.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${LLVM_MAIN_INCLUDE_DIR}/llvm/Bitcode
+  ${LLVM_MAIN_INCLUDE_DIR}/llvm/Bitstream
+  )
