+++ llvm/lib/Bitstream/Reader/LLVMBuild.txt
@@ -1,4 +1,4 @@
-;===- ./lib/Bitcode/Reader/LLVMBuild.txt -----------------------*- Conf -*--===;
+;===- ./lib/Bitstream/Reader/LLVMBuild.txt ---------------------*- Conf -*--===;
 ;
 ; Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 ; See https://llvm.org/LICENSE.txt for license information.
@@ -16,6 +16,6 @@
 
 [component_0]
 type = Library
-name = BitReader
-parent = Bitcode
-required_libraries = BitstreamReader Core Support
+name = BitstreamReader
+parent = Bitstream
+required_libraries = Support
