+++ llvm/lib/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(IRReader)
 add_subdirectory(CodeGen)
 add_subdirectory(BinaryFormat)
 add_subdirectory(Bitcode)
+add_subdirectory(Bitstream)
 add_subdirectory(Transforms)
 add_subdirectory(Linker)
 add_subdirectory(Analysis)
