+++ llvm/unittests/Bitcode/CMakeLists.txt
@@ -8,7 +8,5 @@ set(LLVM_LINK_COMPONENTS
 
 add_llvm_unittest(BitcodeTests
   BitReaderTest.cpp
-  BitstreamReaderTest.cpp
-  BitstreamWriterTest.cpp
   RecordLayout.cpp
   )
